convert link component to typescript
